+++ src/inspect_ai/_view/www/src/state/hooks.ts
@@ -1,5 +1,4 @@
-import { highlightElement } from "prismjs";
-import { RefObject, useCallback, useEffect, useMemo, useRef } from "react";
+import { useCallback, useEffect, useMemo, useRef } from "react";
 import { EvalSample, EvalSpec, Events, Status } from "../@types/log";
 import {
   createEvalDescriptor,
@@ -478,32 +477,6 @@ export const usePrevious = <T>(value: T) => {
   }, [value]);
 
   return ref.current;
-};
-
-// Syntax highlighting strings larger than this is too slow
-const kPrismRenderMaxSize = 250000;
-
-export const usePrismHighlight = (
-  containerRef: RefObject<HTMLDivElement | null>,
-  contentLength: number,
-) => {
-  useEffect(() => {
-    requestAnimationFrame(() => {
-      if (
-        contentLength > 0 &&
-        containerRef.current !== null &&
-        contentLength <= kPrismRenderMaxSize
-      ) {
-        const codeBlocks = containerRef.current?.querySelectorAll("pre code");
-        codeBlocks?.forEach((block) => {
-          if (block.className.includes("language-")) {
-            block.classList.add("sourceCode");
-            highlightElement(block as HTMLElement);
-          }
-        });
-      }
-    });
-  }, [contentLength, containerRef]);
 };
 
 export const useSetSelectedLogIndex = () => {